refactor: move common code into src
