+++ src/common/index.js
@@ -1,4 +1,4 @@
-import WASMAudioDecoderCommon from "./WASMAudioDecoderCommon.js";
-import WASMAudioDecoderWorker from "./WASMAudioDecoderWorker.js";
+import WASMAudioDecoderCommon from "./src/WASMAudioDecoderCommon.js";
+import WASMAudioDecoderWorker from "./src/WASMAudioDecoderWorker.js";
 
 export { WASMAudioDecoderCommon, WASMAudioDecoderWorker };
